Add error codes
In case of errors, all functions that perform business logic now return
error codes instead of -1.
This will make it easier to debug error for clients.

Error codes are in logerrno.h

`log_open` and `segment_open` function signatures have changed to
return error codes instead of NULL. This requires the two functions to
take an additional field (a poiter) which is modified to point to
the newly allocated log_t or segment_t structures.