mbptree: non-leaf nodes do not split
Unlike a standard B+tree, when a new element is inserted into a full leaf
and a new leaf is created, the full leaf is not split and elements do not get
moved over to the new leaf.
This allows to use space more efficiently and reduce allocations of new
nodes. It has been made possible by the data structures' monitonic property.
Non-leaf nodes split as standard B+tree non-leaf nodes.